fix: bump dependencies
We haven't given any love to our gatsby website in a while, and it shows!

This commit fixes the build issues, and bumps all of dependencies.
It also fixes a couple of scss warnings, and checks in package-lock.json so we can be sure gh pages builds the same site as we do locally.

I hope it works! :D